权限

+++ routes/api.js
@@ -9,6 +9,11 @@ router.get('/notes', function(req, res, next){
 })
 
 router.post('/notes/add', function(req, res, next){
+    console.log(req.session.user)
+    if(!req.session.user){
+        return res.send({status:1, errorMsg: '请先登录'})
+    }
+    
     var note = req.body.note
     Note.create({text: note}).then(function(){
         res.send({status: 0})
@@ -19,12 +24,20 @@ router.post('/notes/add', function(req, res, next){
 })
 
 router.post('/notes/edit', function(req, res, next){
+    if(!req.session.user){
+        return res.send({status:1, errorMsg: '请先登录'})
+    }
+
     Note.update({text: req.body.note}, {where:{id:req.body.id}}).then( () => {
         res.send({status: 0})
     })
 })
 
 router.post('/notes/delete', function(req, res, next){
+    if(!req.session.user){
+        return res.send({status:1, errorMsg: '请先登录'})
+    }
+
     Note.destroy({where: {id:req.body.id}}).then( () => {
         res.send({status: 0})
     })
